+++ lib/src/fcm_interface.dart
@@ -1,24 +1,29 @@
+import 'dart:convert';
+import 'dart:html' as html;
 import 'dart:js';
 
 import 'package:firebase_cloud_messaging_interop/src/fcm_interop.dart';
 import 'package:firebase_cloud_messaging_interop/src/js_to_dart_helper.dart';
+import 'package:firebase_cloud_messaging_interop/src/stringify_interop.dart';
 
 /// An interface that make it easier to use FCM interop.
-class FCM {
-  Messaging _messaging;
-
+class FirebaseMessagingWeb {
   /// If the [publicVapidKey] is specified, the init function will be call in
   /// the constructor. If not, then the init method should be called before using
   /// any other function related to [FCM]
-  FCM({String publicVapidKey}) {
-    if (publicVapidKey != null) init(publicVapidKey);
+  FirebaseMessagingWeb({String publicVapidKey}) {
+    if (publicVapidKey != null) {
+      init(publicVapidKey);
+    }
   }
 
-  /// Initiate Firebase Messaging and set the [publicVapidKey].
+  Messaging _messaging;
+
+  /// Initiates Firebase Messaging and set the [publicVapidKey].
   void init(String publicVapidKey) {
     if (publicVapidKey == null || publicVapidKey.isEmpty) {
-      throw FormatException(
-          "The publicVapidKey can not be null nor can it be an empty string !");
+      throw const FormatException(
+          'The publicVapidKey can not be null nor can it be an empty string !');
     }
 
     _messaging = firebase.messaging();
@@ -26,18 +31,29 @@ class FCM {
     _messaging.usePublicVapidKey(publicVapidKey);
   }
 
-  /// Adding a callback allowing to retrieve the new token after it was refreshed.
+  /// Adds a callback allowing to retrieve the new token after it was refreshed.
   void onTokenRefresh(Function() callBack) =>
       _messaging.onTokenRefresh(allowInterop(() => callBack()));
 
-  /// Adding a callback allowing to retrieve the FCM payload
-  void onMessage(Function(Map message) callBack) =>
-      _messaging.onMessage(allowInterop((e) => callBack((toDartMap(e)))));
+  /// Adds a callback allowing to retrieve the FCM payload
+  void onMessage(Function(Map<String, dynamic> message) callBack) {
+    _messaging.onMessage(allowInterop((dynamic jsMessageObject) {
+      return callBack(
+          jsonDecode(stringify(jsMessageObject)) as Map<String, dynamic>);
+    }));
+  }
 
-  /// Delete the current FCM token.
+  /// Prompts the user for notification permissions the first time
+  Future<bool> requestNotificationPermissions() async {
+    final String result = await html.Notification.requestPermission();
+
+    return result == 'granted';
+  }
+
+  /// Deletes the current FCM token.
   Future<bool> deleteToken(String token) =>
       futureFromPromise<bool>(_messaging.deleteToken(token));
 
-  /// Retrieve the client FCM token. (The client must first ask for the permission [Notification.requestPermission])
+  /// Retrieves the client FCM token. (The client must first ask for the permission [requestNotificationPermissions])
   Future<String> getToken() => futureFromPromise<String>(_messaging.getToken());
 }